+++ apps/web/src/entities/notification/model/index.ts
@@ -1,3 +1,2 @@
 export * from './types';
 export * from './queries';
-export * from './hooks';
