+++ apps/web/src/features/notification/model/hooks/usePushNotification.ts
@@ -1,11 +1,12 @@
 import { PushNotifications } from '@capacitor/push-notifications';
 
+import { useUpdateFCMToken } from '@/features/notification';
+
 import { MESSAGE } from '@/shared/constants';
 import { toast } from '@/shared/model';
 import { getNativeFCM, setNativeFCM } from '@/shared/storage';
 import { isWeb } from '@/shared/utils';
 
-import { useUpdateFCMToken } from '../mutations/useUpdateFCMToken';
 //TODO : main에 올리기 전에 console 제거하기
 export const usePushNotification = () => {
   const updateFCMTokenMutation = useUpdateFCMToken();
