+++ apps/web/src/features/notification/model/index.ts
@@ -1,0 +1,2 @@
+export * from './mutations';
+export * from './hooks';
